chore: move components to react folder
