+++ src/app/App.js
@@ -1,6 +1,6 @@
 import 'react-native-gesture-handler';
 import React from 'react';
-import { ThemeProvider } from '../shared/components';
+import { ThemeProvider } from '../shared/react';
 import AppStack, { rootNavigation } from '../navigation';
 import { NavigationContainer } from '@react-navigation/native';
 import { SafeAreaProvider } from 'react-native-safe-area-context';
