+++ src/navigation/profile-stack/header/ProfileHeader.js
@@ -1,7 +1,7 @@
 import React, { useCallback } from 'react';
 import PropTypes from 'prop-types';
 import { View, Text } from 'react-native';
-import { Button, useTheme } from '../../../shared/components';
+import { Button, useTheme } from '../../../shared/react';
 import { SafeAreaView } from 'react-native-safe-area-context';
 
 import buildStyles from './styles';
