+++ src/screens/home/HomeScreen.js
@@ -2,7 +2,7 @@ import React, { useCallback, useMemo } from 'react';
 import PropTypes from 'prop-types';
 import { Text } from 'react-native';
 import HomeHeader from './header';
-import { Button, useTheme } from '../../shared/components';
+import { Button, useTheme } from '../../shared/react';
 import { SafeAreaView } from 'react-native-safe-area-context';
 
 import buildStyles from './styles';
