+++ src/screens/profile/1/ProfileScreen1.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { mount } from 'enzyme';
 import ProfileScreen1 from './ProfileScreen1';
-import { AppTree } from '../../../shared/test-utils/components';
+import { AppTree } from '../../../shared/test-utils/react';
 
 it('should render correctly', () => {
     const tree = mount(
