+++ src/shared/react/index.js
