+++ src/shared/react/index.test.js
