+++ src/shared/react/theme/index.test.js
